+++ src/CesiumWidget/story.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { storiesOf } from "@storybook/react";
 
-import CesiumWidget from "../CesiumWidget";
+import CesiumWidget from "./CesiumWidget";
 
 storiesOf("CesiumWidget", module).add("Basic", () => <CesiumWidget full />);
