+++ src/Entity/story/CanvasEntity.tsx
