+++ src/ImageryLayerCollection/story.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { storiesOf } from "@storybook/react";
 
 import Viewer from "../Viewer";
-import ImageryLayerCollection from "../ImageryLayerCollection";
+import ImageryLayerCollection from "./ImageryLayerCollection";
 import { action } from "@storybook/addon-actions";
 
 storiesOf("ImageryLayerCollection", module).add("Basic", () => (
