+++ src/Label/story.tsx
@@ -1,34 +1,36 @@
 import React from "react";
-import { Cartesian3, Transforms } from "cesium";
+import { Cartesian3, Color, Transforms } from "cesium";
 import { storiesOf } from "@storybook/react";
 import { action } from "@storybook/addon-actions";
 
 import Viewer from "../Viewer";
-import Polyline from "../Polyline";
-import PolylineCollection from "../PolylineCollection";
-import CameraFlyTo from "../CameraFlyTo";
+import Label from "./Label";
+import LabelCollection from "../LabelCollection";
 
 const center = Cartesian3.fromDegrees(-75.59777, 40.03883);
-const positions = [
-  new Cartesian3(-75, 35, 0),
-  new Cartesian3(-125, 35, 0),
-  new Cartesian3(-125, 135, 0),
-];
-storiesOf("Polyline", module)
+
+storiesOf("Label", module)
   .add("Basic", () => (
     <Viewer full>
-      <PolylineCollection modelMatrix={Transforms.eastNorthUpToFixedFrame(center)}>
-        <Polyline positions={positions} width={10} />
-      </PolylineCollection>
-      <CameraFlyTo duration={0} destination={Cartesian3.fromDegrees(-75.6, 40.04, 1000)} />
+      <LabelCollection modelMatrix={Transforms.eastNorthUpToFixedFrame(center)}>
+        <Label fillColor={Color.ORANGE} position={new Cartesian3(0.0, 0.0, 0.0)} text="Cesium" />
+        <Label
+          fillColor={Color.YELLOW}
+          position={new Cartesian3(1000000.0, 0.0, 0.0)}
+          text="resium"
+        />
+        <Label fillColor={Color.GREEN} position={new Cartesian3(0.0, 1000000.0, 0.0)} text="Hi!" />
+        <Label fillColor={Color.CYAN} position={new Cartesian3(0.0, 0.0, 1000000.0)} text="LGTM" />
+      </LabelCollection>
     </Viewer>
   ))
   .add("Events", () => (
     <Viewer full>
-      <PolylineCollection modelMatrix={Transforms.eastNorthUpToFixedFrame(center)}>
-        <Polyline
-          positions={positions}
-          width={10}
+      <LabelCollection modelMatrix={Transforms.eastNorthUpToFixedFrame(center)}>
+        <Label
+          fillColor={Color.ORANGE}
+          position={new Cartesian3(0.0, 0.0, 0.0)}
+          text="Cesium"
           onClick={action("onClick")}
           onDoubleClick={action("onDoubleClick")}
           onMouseDown={action("onMouseDown")}
@@ -47,7 +49,6 @@ storiesOf("Polyline", module)
           onMouseEnter={action("onMouseEnter")}
           onMouseLeave={action("onMouseLeave")}
         />
-      </PolylineCollection>
-      <CameraFlyTo duration={0} destination={Cartesian3.fromDegrees(-75.6, 40.04, 1000)} />
+      </LabelCollection>
     </Viewer>
   ));
