+++ src/PointPrimitive/story.tsx
@@ -4,7 +4,7 @@ import { storiesOf } from "@storybook/react";
 import { action } from "@storybook/addon-actions";
 
 import Viewer from "../Viewer";
-import PointPrimitive from "../PointPrimitive";
+import PointPrimitive from "./PointPrimitive";
 import PointPrimitiveCollection from "../PointPrimitiveCollection";
 
 const center = Cartesian3.fromDegrees(-75.59777, 40.03883);
