+++ src/Polyline/story.tsx
@@ -1,36 +1,34 @@
 import React from "react";
-import { Cartesian3, Color, Transforms } from "cesium";
+import { Cartesian3, Transforms } from "cesium";
 import { storiesOf } from "@storybook/react";
 import { action } from "@storybook/addon-actions";
 
 import Viewer from "../Viewer";
-import Label from "../Label";
-import LabelCollection from "../LabelCollection";
+import Polyline from "./Polyline";
+import PolylineCollection from "../PolylineCollection";
+import CameraFlyTo from "../CameraFlyTo";
 
 const center = Cartesian3.fromDegrees(-75.59777, 40.03883);
-
-storiesOf("Label", module)
+const positions = [
+  new Cartesian3(-75, 35, 0),
+  new Cartesian3(-125, 35, 0),
+  new Cartesian3(-125, 135, 0),
+];
+storiesOf("Polyline", module)
   .add("Basic", () => (
     <Viewer full>
-      <LabelCollection modelMatrix={Transforms.eastNorthUpToFixedFrame(center)}>
-        <Label fillColor={Color.ORANGE} position={new Cartesian3(0.0, 0.0, 0.0)} text="Cesium" />
-        <Label
-          fillColor={Color.YELLOW}
-          position={new Cartesian3(1000000.0, 0.0, 0.0)}
-          text="resium"
-        />
-        <Label fillColor={Color.GREEN} position={new Cartesian3(0.0, 1000000.0, 0.0)} text="Hi!" />
-        <Label fillColor={Color.CYAN} position={new Cartesian3(0.0, 0.0, 1000000.0)} text="LGTM" />
-      </LabelCollection>
+      <PolylineCollection modelMatrix={Transforms.eastNorthUpToFixedFrame(center)}>
+        <Polyline positions={positions} width={10} />
+      </PolylineCollection>
+      <CameraFlyTo duration={0} destination={Cartesian3.fromDegrees(-75.6, 40.04, 1000)} />
     </Viewer>
   ))
   .add("Events", () => (
     <Viewer full>
-      <LabelCollection modelMatrix={Transforms.eastNorthUpToFixedFrame(center)}>
-        <Label
-          fillColor={Color.ORANGE}
-          position={new Cartesian3(0.0, 0.0, 0.0)}
-          text="Cesium"
+      <PolylineCollection modelMatrix={Transforms.eastNorthUpToFixedFrame(center)}>
+        <Polyline
+          positions={positions}
+          width={10}
           onClick={action("onClick")}
           onDoubleClick={action("onDoubleClick")}
           onMouseDown={action("onMouseDown")}
@@ -49,6 +47,7 @@ storiesOf("Label", module)
           onMouseEnter={action("onMouseEnter")}
           onMouseLeave={action("onMouseLeave")}
         />
-      </LabelCollection>
+      </PolylineCollection>
+      <CameraFlyTo duration={0} destination={Cartesian3.fromDegrees(-75.6, 40.04, 1000)} />
     </Viewer>
   ));
