+++ src/Scene/story.tsx
@@ -3,7 +3,7 @@ import { SceneMode } from "cesium";
 import { storiesOf } from "@storybook/react";
 
 import Viewer from "../Viewer";
-import Scene from "../Scene";
+import Scene from "./Scene";
 
 storiesOf("Scene", module).add("Basic", () => (
   <Viewer full>
