+++ src/ScreenSpaceCameraController/story.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { storiesOf } from "@storybook/react";
 
 import Viewer from "../Viewer";
-import ScreenSpaceCameraController from "../ScreenSpaceCameraController";
+import ScreenSpaceCameraController from "./ScreenSpaceCameraController";
 
 storiesOf("ScreenSpaceCameraController", module).add("Basic", () => (
   <Viewer full>
