+++ src/Sun/story.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { storiesOf } from "@storybook/react";
 
 import Viewer from "../Viewer";
-import Sun from "../Sun";
+import Sun from "./Sun";
 
 storiesOf("Sun", module).add("Basic", () => (
   <Viewer full>
