+++ src/Viewer/story.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { storiesOf } from "@storybook/react";
 
-import Viewer from "../Viewer";
+import Viewer from "./Viewer";
 
 storiesOf("Viewer", module).add("Basic", () => <Viewer full />);
